Tests: Remove `originalContent` from fixtures
Previously we've been asserting in our full-content snapshot tests that
the `originalContent` property doesn't change, but it's faulty and going
away (see #38460).

By removing it from the test fixtures we can relax our constraint that
it exists and doesn't change, leaving future work easier to review and focus.

This is a generated property while loading the blocks and use of it leads
to content corruption and loss on certain invalid blocks and blocks with
inner blocks.

After this change we're still veryifying the full load and save flow
for the block implementations.